Online Bookstore - "The Bookstore"
This commits introduces the online Bookstore called "The Bookstore"
The application was written in C# with using ASP.NET technology and architecture pattern MVC 5.
There are three main parts to the application: shopping, checkout, and administration.

Shopping:
   -> visitors can browse books by genre,
   -> view a single book and add it to their cart,
   -> review their cart, removing any items they no longer want.

Checkout :
Checkout will prompt them to login or register for a user account, after creating an account, they can complete the order by filling out shipping and payment information. The completed transactions ends with simple confirmation screen.

Administration :
In addition to user interface, there is an administrator section, which allows to:
   -> add booksto store
   -> remove books from store
   -> edit existing books in store

The database was created using Entity Framework with Code First approach.